Common: Move types/plugin_types.py classes to agent_plugin_manifest.py
Define things close to where they're used.
