+++ monkey/common/agent_plugins/agent_plugin_manifest.py
@@ -1,11 +1,41 @@
+import re
 from typing import Callable, Mapping, Optional, Tuple, Type
 
-from pydantic import HttpUrl
+from pydantic import ConstrainedStr, HttpUrl
+from semver import VersionInfo
 
 from common import OperatingSystem
 from common.agent_plugins import AgentPluginType
 from common.base_models import InfectionMonkeyBaseModel, InfectionMonkeyModelConfig
-from common.types import PluginName, PluginVersion
+
+
+class PluginName(ConstrainedStr):
+    """
+    A plugin name
+
+    Allowed characters are alphanumerics and underscore.
+    """
+
+    strip_whitespace = True
+    regex = re.compile("^[a-zA-Z0-9_]+$")
+
+
+class PluginVersion(VersionInfo):
+    @classmethod
+    def __get_validators__(cls):
+        """Return a list of validator methods for pydantic models."""
+        yield cls.parse
+
+    @classmethod
+    def __modify_schema__(cls, field_schema):
+        """Inject/mutate the pydantic field schema in-place."""
+        field_schema.update(
+            examples=[
+                "1.0.2",
+                "2.15.3-alpha",
+                "21.3.15-beta+12345",
+            ]
+        )
 
 
 class AgentPluginManifest(InfectionMonkeyBaseModel):
