+++ monkey/common/types/__init__.py
@@ -3,7 +3,5 @@ from .serialization import JSONSerializable
 from .ids import AgentID, HardwareID, MachineID
 from .int_range import IntRange
 from .networking import NetworkService, NetworkPort, PortStatus, SocketAddress, NetworkProtocol
-from .plugin_types import PluginName
-from .plugin_types import PluginVersion
 from .secrets import OTP, Token
 from .file_extension import FileExtension
